float

+++ Engine/game/scene/gameScene/enemy/Enemy.cpp
@@ -39,7 +39,7 @@ void Enemy::Collision()
 	Vector2 mapCenter = map_->GetSprite()->GetPosition();
 	Vector2 enemyPos = sprite_->GetPosition();
 
-	float length = std::sqrt(std::pow(enemyPos.x - mapCenter.x, 2) + std::pow(enemyPos.y - mapCenter.y, 2));
+	float length = std::sqrt(std::powf(enemyPos.x - mapCenter.x, 2) + std::powf(enemyPos.y - mapCenter.y, 2));
 
 	// length以下なら当たっている
 	if (length < (map_->GetSprite()->GetSize().x / 2.0f)) {
